+++ sistema-certificados-backend/src/main/java/com/agg/certificados/dtos/response/DataGeneratorResponseDto.java
@@ -18,7 +18,7 @@ public class DataGeneratorResponseDto {
     public String address_rcd;
     public String reception_date_rcd;
     public TypeWeightResponseDto type_weight;
-    public Long total_rcd;
+    public double total_rcd;
     public List<QuantitiesRcdResponseDto> quantitiesRcd;
     public DataDriverResponseDto data_driver;
     public CertificationMiniResponseDto certification;
